+++ docs-markdown/changelog.md
@@ -7,6 +7,8 @@
 - Collapse relative link improvements
 - Remote-WSL bug fix: [Issue 605](https://github.com/microsoft/vscode-docs-authoring/issues/605)
 - Added Unit Test Coverage: table-controller.ts
+- Single value cleanup bug fix: [Issue 252](https://github.com/microsoft/vscode-docs-authoring/issues/252)
+- Bug fix for cleanup metadata removing body text
 
 ## 0.2.73 (June 8th, 2020)